requirements.txt
fastapi
uvicorn
py-eureka-client
pymongo
psycopg2-binary
apscheduler

#pip install -r requirements.txt
#uvicorn main:app --reload --host 0.0.0.0 --port 8001